First try about Payment implementation
Not tested, just coded

+++ FoodYeah/Dto/LOCDto.cs
@@ -7,23 +7,23 @@ namespace FoodYeah.Dto
 {
     public class LOCDto
     {
-        public float TCEA { get; set; }
-        public int LineOfCredit { get; set; }
-        public int AvalibleLineOfCredit { get; set; }
+        public decimal TCEA { get; set; }
+        public decimal LineOfCredit { get; set; }
+        public decimal AvalibleLineOfCredit { get; set; }
 
         
     }
     public class CreateLOCDto
     {
         public int CustomerId { get; set; }
-        public float TCEA { get; set; }
-        public int TotalLineOfCredit { get; set; }
+        public decimal TCEA { get; set; }
+        public decimal TotalLineOfCredit { get; set; }
         
     }
     public class UpdateLOCDto
     {
-        public float TCEA { get; set; }
-        public int TotalLineOfCredit { get; set; }
+        public decimal TCEA { get; set; }
+        public decimal TotalLineOfCredit { get; set; }
        
     }
     public class DeleteLOCDto
